Update status and fix several bugs
